Resolve project structure
