+++ cultr/app.py
@@ -2,12 +2,13 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 from .database import init_database
-from .routers import oauth2, urls
+from .routers import login, users, urls
 from .config import settings
 
 app = FastAPI()
-app.include_router(oauth2.router, prefix="/api", tags=["oauth"])
-app.include_router(urls.api_router, prefix="/api/v1", tags=["urls"])
+app.include_router(login.router, prefix="/api", tags=["login"])
+app.include_router(users.router, prefix="/api/users", tags=["users"])
+app.include_router(urls.api_router, prefix="/api/v1/urls", tags=["urls"])
 app.include_router(urls.redirect_router, tags=["urls"])
 
 if settings.CORS_ORIGINS:
